requirements.txt
azure-identity==1.21.0